Share renderer logic across web and Electron
