+++ src/index.html
@@ -76,6 +76,6 @@
     <div id="info-details"></div>
   </div>
 
-  <script src="renderer.js"></script>
+  <script type="module" src="renderer.js"></script>
 </body>
 </html>
